cs/requirements.txt
pandas
sentence-transformers
faiss-cpu
tqdm
requests
numpy
glob2
